Renamed model file to 'roomModel'.
